chat_db/requirements.txt
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.16
Flask-Script==2.0.6
py-ms>=1.0.2